added support for timer with callback to stop timing
extracted base class from TimerMetric that encapsulates the common Timer functionality.
Kept the Time(Action) methods in TimerMetric and added CallbackTimerMetric with a
Context Time(); method. The timer starts when the Time() method is called and is measured when the Stop() method is called on the returned context object.

This is in-line with the original codahale metrics Timer mechanism.